Convert template class scan_codec_factory into template method
The class scan_codec_factory has no state and can be replaced by just a template function.
