+++ spelling.dic
@@ -1,0 +1,13 @@
+jpegls
+cppcoreguidelines
+golomb
+charls
+opto
+Errval
+Golomb
+maxval
+unmap
+subspan
+misc
+bugprone
+countl
