+++ src/CMakeLists.txt
@@ -121,13 +121,13 @@ target_sources(charls
     "${CMAKE_CURRENT_LIST_DIR}/jpegls_preset_coding_parameters.h"
     "${CMAKE_CURRENT_LIST_DIR}/jpegls_preset_parameters_type.h"
     "${CMAKE_CURRENT_LIST_DIR}/lossless_traits.h"
+    "${CMAKE_CURRENT_LIST_DIR}/make_scan_codec.h"
+    "${CMAKE_CURRENT_LIST_DIR}/make_scan_codec.cpp"
     "${CMAKE_CURRENT_LIST_DIR}/process_decoded_line.h"
     "${CMAKE_CURRENT_LIST_DIR}/process_encoded_line.h"
     "${CMAKE_CURRENT_LIST_DIR}/quantization_lut.h"
     "${CMAKE_CURRENT_LIST_DIR}/quantization_lut.cpp"
     "${CMAKE_CURRENT_LIST_DIR}/scan_codec.h"
-    "${CMAKE_CURRENT_LIST_DIR}/scan_codec_factory.h"
-    "${CMAKE_CURRENT_LIST_DIR}/scan_codec_factory.cpp"
     "${CMAKE_CURRENT_LIST_DIR}/scan_decoder.h"
     "${CMAKE_CURRENT_LIST_DIR}/scan_decoder_impl.h"
     "${CMAKE_CURRENT_LIST_DIR}/scan_encoder.h"
