+++ src/CharLS.vcxproj
@@ -216,7 +216,7 @@
     <ClCompile Include="validate_spiff_header.cpp" />
     <ClCompile Include="version.cpp" />
     <ClCompile Include="charls_jpegls_decoder.cpp" />
-    <ClCompile Include="scan_codec_factory.cpp" />
+    <ClCompile Include="make_scan_codec.cpp" />
     <ClCompile Include="charls_jpegls_encoder.cpp" />
     <ClCompile Include="jpegls_error.cpp" />
     <ClCompile Include="jpeg_stream_reader.cpp" />
@@ -243,7 +243,7 @@
     <ClInclude Include="default_traits.h" />
     <ClInclude Include="scan_encoder.h" />
     <ClInclude Include="golomb_lut.h" />
-    <ClInclude Include="scan_codec_factory.h" />
+    <ClInclude Include="make_scan_codec.h" />
     <ClInclude Include="jpegls_algorithm.h" />
     <ClInclude Include="jpegls_preset_coding_parameters.h" />
     <ClInclude Include="jpeg_marker_code.h" />
