+++ src/CharLS.vcxproj.filters
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup>
-    <ClCompile Include="scan_codec_factory.cpp">
+    <ClCompile Include="make_scan_codec.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="jpegls_error.cpp">
@@ -69,7 +69,7 @@
     <ClInclude Include="scan_encoder.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="scan_codec_factory.h">
+    <ClInclude Include="make_scan_codec.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="jpeg_marker_code.h">
